+++ examples/lib/stories/widgets/sprite_widget.dart
@@ -1,5 +1,4 @@
 import 'package:dashbook/dashbook.dart';
-import 'package:flame/flame.dart';
 import 'package:flame/widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
@@ -11,8 +10,8 @@ Widget spriteWidgetBuilder(DashbookContext ctx) {
     width: ctx.numberProperty('container width', 400),
     height: ctx.numberProperty('container height', 200),
     decoration: BoxDecoration(border: Border.all(color: Colors.amber)),
-    child: SpriteWidget(
-      sprite: Sprite(Flame.images.fromCache('shield.png')),
+    child: SpriteWidget.asset(
+      path: 'shield.png',
       anchor: Anchor.valueOf(
         ctx.listProperty('anchor', 'center', anchorOptions),
       ),